Year1/BSMS1306/requirements.txt
matplotlib
streamlit
pandas
plotly.express
seaborn
